+++ src/components/molecules/Table/HoverTableCell.module.scss
@@ -1,0 +1,28 @@
+@use "src/styles/libs" as *;
+
+.table-cell-ellipsis {
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+.table-cell-hover {
+  position: absolute;
+  top: 7px;
+  left: 8px;
+  z-index: 10;
+  display: flex;
+
+  width: fit-content;
+  padding: 0px 7px;
+  border: 1px solid $BLUISH-GRAY-200;
+  border-radius: 4px;
+
+  background-color: $WHITE;
+
+  white-space: pre;
+
+  transition: background-color 0.3s, border 0.3s;
+
+  @include fade-in(0.2s);
+}
