+++ src/components/molecules/Table/TableCell.tsx
@@ -1,13 +1,7 @@
-import {
-  CSSProperties,
-  PropsWithChildren,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
-import styles from "./index.module.scss";
+import { CSSProperties, PropsWithChildren } from "react";
 import { cleanClassName } from "@/utils";
 import { isString } from "@/utils/public/is";
+import { HoverTableCell } from "@/components/molecules/Table/HoverTableCell";
 export interface TableCellProps extends PropsWithChildren {
   className?: string;
   maxWidth?: CSSProperties["maxWidth"];
@@ -22,58 +16,14 @@ export const TableCell = ({
   return (
     <td className={`${cleanClassName(classes)} `}>
       {maxWidth ? (
-        <LongTextCell
+        <HoverTableCell
           maxWidth={isString(maxWidth) ? parseInt(maxWidth, 10) : maxWidth}
         >
           {children}
-        </LongTextCell>
+        </HoverTableCell>
       ) : (
         <div>{children}</div>
       )}
     </td>
-  );
-};
-
-interface LongTextCellProps extends PropsWithChildren {
-  maxWidth: number;
-}
-const LongTextCell = ({ children, maxWidth }: LongTextCellProps) => {
-  const [isHover, setIsHover] = useState(false);
-  const [isTextTooShort, setIsTextTooShort] = useState(false);
-  const divRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (divRef.current) {
-      const contentCell = divRef.current;
-      const isNotElipsissed = contentCell.scrollWidth > contentCell.clientWidth;
-
-      if (isNotElipsissed) {
-        setIsTextTooShort(true);
-      }
-    }
-  }, []);
-
-  return (
-    <>
-      {isTextTooShort && isHover ? (
-        <div
-          onMouseLeave={() => isTextTooShort && setIsHover(false)}
-          className={styles["long-text"]}
-          data-testid="j"
-        >
-          {children}
-        </div>
-      ) : (
-        <div
-          ref={divRef}
-          data-testid="j2"
-          style={{ width: maxWidth }}
-          onMouseEnter={() => isTextTooShort && setIsHover(true)}
-          className={`${styles["table-cell-ellipsis"]}`}
-        >
-          {children}
-        </div>
-      )}
-    </>
   );
 };
