+++ src/components/molecules/Table/index.module.scss
@@ -12,30 +12,3 @@ td {
 td {
   position: relative;
 }
-
-.table-cell-ellipsis {
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.long-text {
-  position: absolute;
-  top: 7px;
-  left: 8px;
-  z-index: 10;
-  display: flex;
-
-  width: fit-content;
-  padding: 0px 7px;
-  border: 1px solid $BLUISH-GRAY-200;
-  border-radius: 4px;
-
-  background-color: $WHITE;
-
-  white-space: pre;
-
-  transition: background-color 0.3s, border 0.3s;
-
-  @include fade-in(0.2s);
-}
