+++ src/utils/private/dummy.ts
@@ -85,3 +85,347 @@ export const getCustomDateString = (daysAgo: number) => {
 
   return `${year}-${month}-${day}T${hours}:${minutes}:${seconds}${timezoneOffsetString}`;
 };
+
+// eslint-disable-next-line storybook/prefer-pascal-case
+export const tableData = {
+  titles: [
+    "학교",
+    "이름",
+    "학번",
+    "학점",
+    "군필",
+    "졸업여부",
+    "자기소개",
+    "복사가능한 자기소개",
+  ],
+  bodies: [
+    {
+      학교: "충북대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: "안녕하세요 ",
+      복사가능한_자기소개: "안녕하세요 copy",
+    },
+    {
+      학교: "서울대학교",
+      이름: "류준열2",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `안녕하세요 여러분 이건 매우 긴 텍스트입니다. 
+        처음에는 말줄임표로 나타나지만, 
+        Hover하면 전체 내용을 볼 수 있답니다.`,
+      복사가능한_자기소개: `안녕하세요 여러분 이건 매우 긴 텍스트입니다. 
+        처음에는 말줄임표로 나타나지만, 
+        Hover하면 전체 내용을 볼 수 있답니다.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+    {
+      학교: "하버드대학교",
+      이름: "류준열",
+      학번: "2015039077",
+      학점: 4.5,
+      군필: true,
+      졸업여부: true,
+      자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+      복사가능한_자기소개: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+Aenean nec mollis nulla. 
+Phasellus lacinia tempus mauris eu laoreet. 
+Proin gravida velit dictum dui consequat malesuada.`,
+    },
+  ],
+};
